+++ Albatross.Hosting/ExceptionHandling/ProblemDetailsWithTraceId.cs
@@ -1,7 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-
-namespace Albatross.Hosting.ExceptionHandling {
-	public class ProblemDetailsWithTraceId : ProblemDetails{
-		public string? TraceId { get; set; }
-	}
-}